refactor ocr as CamItem
